requirements.txt
psycopg2-binary
boto3==1.26.22